initial commit for automated log uploader
This repository contains scripts designed to upload logs to gw2raidar and
dps.report, and then format them into a nice message to a discord webhook.

It is pretty customized and may not be of much use, but could provide good
inspiration for developing your own tools.

At the very least, it functions as a good example for how to use
PowerShell to do Rest API calls, and how to hook into the gw2raidar API as
well as the dps.report API.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>