switch to using a configuration file
Instead of storing the customizable variables within the scripts
themselves, store them in a simple JSON formatted file. You may need to
adjust the programs to point to the configuration file depending on how
you invoke the scripts.

This new method also enables us to add a few more features (such as
debug_mode), and additionally makes it easier to keep the scripts up to
date.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>