configure powershell to immediately terminate on error
In general, the script errors will occur due to things like typos in the
configuration file, or files that are out of place. Currently, the
powershell scripts will continue, which can cause problems.

A bandaid fix is to set the ErrorActionPreference variable to "Stop" so
that we at least terminate immediately upon error. This should help
prevent cascading issues from occurring.

Future plans include catching and handling more errors to provide a better
user experience.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>