upload-logs: regenerate ancillary data
In case the folder got created but some data did not get built properly,
delete the folder and regenerate it. We do this because the data held
there should have only been generated by this script, and it prevents
issues due to previous misconfiguration which caused the data to be
generated incorrectly.

Without this change, it is likely that early misconfiguration that is
fixed will still result in issues later on.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>