further enhancements to error checking
Check for even more files using X-Test-Path, and make sure that the
expected locations exist. Give users more explicit messages than the
exceptions generated when these are missing.

Additionally, create the discord JSON data directory if it doesn't exist
yet, and warn if the path to the last_format_time file is not valid.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>