upload-logs: store start time and gw2raidar response
In case of future API updates, store the gw2raidar JSON response. This
normally includes the upload_id and the filename that we uploaded. This
may be useful in the future if the API is extended to allow requesting
status of a single upload_id.

Additionally, store the start time in the upload extras data folder as
well as the start map.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>