migrate config parsing code to the shared module
Add a new Load-Configuration function which loads and validates the
configuration file before returning a config object.

This replaces a bunch of code shared between multiple scripts, and makes a
more robust configuration parser.

When parsing the configuration file, exclude any unknown fields, and
verify that all fields match the expected type.

This will ensure that accidentally including new fields will not pollute
the config space.

Check for a new config_version field which we expect to be set to 1. This
will be used in future updates which break the API of the configuration
file. For now, we do not require the field. In a future update we may
begin requiring it.

The end result of this is simpler script code, and better error handling
of the configuration file in general.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>