simplify handling of gw2raidar challenge motes
The gw2raidar website splits the challenge mote encounters apart, but for
now, we just want to handle these as the same type of encounter as the
normal runs.

It turns out that the gw2raidar CM ids are just the regular ids with 8
higher bits set.

Instead of storing two separate mappings, just use bitwise arithmetic to
extract the lower 16 bits instead when parsing the id.

This vastly simplifies our handling of the raid challenge motes, as we'll
correctly treat them the same without needing to generate a cm_id and
normal id.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>