add support for showing fractal CM bosses
Now that we can correctly parse the fractal challenge name and Id, use
this data to allow the script to publish these encounters.

We probably want to find a way to separate these from normal encounters,
possibly finding some way based on players, or time, so that we split sets
of encounters more easily.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>