simpleArcParse: update simpleArcParse and add version
Update the simpleArcParse code, and add a version check type to indicate
what version of the binary is currently being used. We'll call the first
version "v0.10" in preparation for the upcoming v0.10 release of
L0G-101086.

The primary change is to use a data structure to store parsed data,
instead of keeping it in local variables ad-hoc. This will be used by
future extensions to further the type of data that we extract.

Update the coding style to avoid using CamelCase in the function and
variable names.

Add doc comments for each function explaining their use.

For combat events, add a parser typedef which we'll use to pass function
pointers to determine what gets parsed. This helps avoid the need to scan
*every* combat event if possible.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>