extract configuration validation to separate function
In preparation for handling a new configuration format, extract the valid
fields for a v1 config object to a global variable. Use a separate
function to perform the configuration validation.

This paves the way to add a v2ConfigurationFields variable which contains
the valid configuration for a v2 config file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>