assume an empty config_version is '1'
Since many users likely have configuration files written prior to the
addition of a config_version field, ensure that the verison is set to 1 if
it is empty. This fixes a minor bug when comparing versions if the
config_version isn't set, and additionally will be important once
configuration format v2 is available.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>