store $config_file variable for configuration scripts
These scripts were written with a $config_file variable in place, but this
was inadvertently removed when we migrated configuration to the shared
powershell module. Now, the backup file is incorrect, and config file is
not copied properly.

Fix this by restoring the $config_file variable we had used previously.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>