use array instead of hash for configuration fields
The order of obtaining hash keys is not deterministic. This causes
Select-Object to randomly order the note properties of the object it
returns.

This would be fine in most circumstances, but in the case where we write
this configuration back to disk, it would be nice if we maintained a
static order.

If we switch to using an array, then the order is guaranteed. Do so, and
order the parameters in a way that makes logical sense, with
config_version and debug_mode at the top, followed by the path fields,
followed by the API tokens, and finally the guild related data.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>