add support for opening a v2 configuration file
Now that we have a script to migrate the v1 configuration file to the v2
format, we need to be able to validate the new format.

Since the guilds array is going to have many sub fields, we should
validate the configuration recursively.

Extract object validation from the main Validate-Configuration function.
This new function, Validate-Object-Fields, is given the array of field
definitions directly. Teach it to understand the validFields and
arrayFields parameters.

These parameters cause the function to recursively call it self on a
subobject. In the case of validFields, treat the field as an object and
validate its immediate fields.

In the case of arrayFields, treat the field as an array, and validate each
object in the array against the arrayFields definition.

For now, the recursive call requires every parameter in the suboject to
exist. Potentially we could expand this to allow each field to define its
set of required and optional fields. This is not yet necessary for the
guilds subfield, so it will be left as a future exercise.

The error messaging when fields are not valid could use some work, but it
does at least correctly validate the format.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>