use correct RequiredFields variable
We accidentally used RequiredParameters, instead of RequiredFields when
parsing required fields. This sort of worked because we happend to have a
global RequiredParameters variable. Unfortunately this broke once we
started validating subfields of the config object.

Fix this so that we don't attempt to check for required fields which
aren't part of the sub object.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>