implement Determine-Guild function
Attempt to determine which guild an encounter belongs to. To do so, use
the discord map for the list of player accounts which are considered guild
members.

A guild will be the "hosting" guild if it has the most players involved in
the encounter. Ties are broken via the priority number.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>