support configuration v2 for the upload-logs.ps1 script
Now that we can determine which guild hosted an encounter, update the
upload-logs.ps1 script to support the new format.

This new version sees fairly minor changes. First, we need to determine
the guild for the encounter, and store it in the JSON. If no guild is
found, we'll throw an exception to skip the upload entirely. This is
because we need the guild data in order to format encounters.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>