migrate format-encounters to config version2
Now that we've migrated the upload-logs.ps1 script, we have guild data
associated with each encounter.

This allows us to associate encounters with guilds, thus enabling posting
different bosses to different discord channels.

To support this, we need to significantly refactor the format script to
handle this. We can mostly keep the formatting code the same by wrapping
it in a function.

While we're doing this, separate out the fractals from the raid bosses,
so that they are published in separate sections. This helps in the
readability of the posts, especially when concerning the player list.

It also allows us to taylor the fractal display to show "fractals" instead
of "wings".

Ultimately, we now support uploading encounters to separate guild discord
channels depending on which players participated in the encounter.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>