format-encounters: use local evtc_name variable
We incorrectly tried to combine the guild data using the $boss.evtc
variable, which isn't set yet. Fix this by using the correct local
$evtc_name variable.

This should fix an issue where the gw2raidar_tag was not configured. We
happened to fall back to the gw2raidar tag as a way of determining the
guild, which works if it is configured properly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>