upload-logs: warn about outdated ArcDPS version
gw2raidar is very picky about what encounters it accepts, and will only
publish encounters which are created with the most recent ArcDPS version.
This can confuse users who happen to forget to update, or are attempting
to upload previous encounters. Add a warning about this in the log
message, by comparing the most recent ArcDPS release with the version in
each file.

We still attempt to upload even if the encounter is out of date, in the
off chance that we are incorrect, and gw2raidar will accept the encounter
after all.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>