upload-logs: fix typo in evtc_arcdps_version variable
When introducing the outdated ArcDPS version check, I accidentally spelled
the variable evtc_arcpds_version for some instances. This caused the one
correct spelling to result in a $null pointer, which caused the try block
to execute inadvertently.

Fix these to use the correctly named evtc_arcdps_version. This will allow
the warning about outdateda arcpds version to be printed.

Fixes: 8a14b9e ("upload-logs: warn about outdated ArcDPS version", 2018-06-06)
Signed-off-by: Jacob Keller <jacob.keller@gmail.com>