config: allow marking subfields as optional
Allow marking object subfields as optional, so that we can introduce a new
field to the guild configuration without breaking backwards compatibility.

If a configuration field is marked as optional, then it will not be passed
to the list of required paramters when validating a sub object. All fields
not marked as "optional" will be required.

For now, leave all current fields as non-optional. A future patch will
introduce the first optional field for guild configurations.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>