add new "everything" config option to guilds
Allow marking a guild with the "everything" option as a new boolean
option. If set to true, then all encounters will be published to this
guild in addition to the normal guild rules.

Multiple guilds may be marked as "everything", and each one will receive
every publish.

Guild disambiguation is handled by the discord webhook URL, so multiple
guilds using the same webhook URL will not cause repeat posts to that
discord.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>