fix completion trigger for gw2raidar scanning
The logic for when we exit the gw2raidar polling is incorrect. In order to
decide if we're complete, we loop over all the guilds, and see if there
are any spaces left for encounters in the boss list.

In response to this, the complete variable was incorrectly set to $false,
rather than $true. Even worse, this check happens after the normal check
for whether gw2raidar has any encounters left.

For the case where we did loop through the entire list of encounters from
gw2raidar and set complete to true, but also managed to *not* be missing
any encounters, this resulted in an infinite loop over every encounter.

This doesn't trigger for many users because they happen to never trigger
the logic for having no missing encounters. This is even further likely to
happen if the user has configured multiple guilds.

Fix this by setting $continue to $true instead of $false, which will
correctly stop the script when we no longer have any encounters missing.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>