support .evtc.zip files with .tmp files inside
In some cases, the evtc.zip files will have their uncompressed file name
suffixed with the .tmp extension. It is believed, but not confirmed, to be
caused by newer experimental versions of ArcDPS.

Allow these compressed zips to function by checking for the .tmp file if
the normal file doesn.t exist.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>