extract logic for formatting the discord webhook embed
In preparation for enabling the ability to publish the same set of bosses
to multiple discord webhook URLs, extract the logic for building up the
webhook embed to its own function.

Since the future extension is going to require using a separate webhook
URL, and a separate set of emojis, we'll also pass those separately
instead of using the ones found inside the $guild object.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>