format-encounters: add experimental support for wing6 bosses
Add the wing6 bosses to the boss list. Because gw2raidar doesn't list the
wing6 boss ids in it's areas API, we need to hack these into the array
manually.

You likely will not save evtc files for these new encounters without using
the experimental development build of ArcDPS. Additionally, currently the
encounters do not match properly.

This is still experimental and is likely to change in the near future.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>