module: add support for converting to and from %UserProfile%
Instead of always converting the config from %UserProfile% on load, allow
specifying whether to convert a $config object to or from %UserProfile%
notation.

This will be used by a later patch which adds support for writing out the
configuration file, in order to convert back to %UserProfile% before
dumping the config to disk.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>