add function to write configuration to disk
Writing the configuration back to disk is problematic, because we convert
%UserProfile% to the environment variable. However, if we just dump this
back to JSON it is confusing to users.

Instead, add a new Write-Configuration function which will re-validate the
configuration and convert the UserProfile paths back to %UserProfile% as
expected.

This will be used by scripts which modify the configuration.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>