simpleArcParse: fix parse_all_cbt_events function
The parse_all_cbt_events function failed to use the parser index variable,
so it simply re-tried the first parser over and over again. It so happens
that we do not actually use the parse_all_cbt_events function currently,
so no real functionality was broken.

Fix this by indexing into the parsers array properly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>