simpleArcParse: prepare for revision 1 support
arcdps recently updated the EVTC file format to include a revision
indicator in place of a previously always NUL byte.

This revision indicator is used as a marker to determine the layout of the
evtc_cbtevent structure.

We use this structured data to parse out details about the events, in
order to find the start and end times.

In order to support the revision 1 format of newly created logs, we need
to be able to change the evtc_cbtevent structure. However, we do not want
to arbitrarily drop support for the old v0 stuff.

Instead, prepare the code base to allow handling both v0 and v1. Use C++
templates, in order to reduce the code  duplication that could result.

For now, we just modify the parse_* functions to be templates taking the
Event type equal to evtc_cbtevent.

A future patch will actually add the new evtc_cbtevent layout and add
support for the v1 combat events.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>