simpleArcParse: add support for parsing new .evtc files
The new version of arcdps updated the combat event layout and modified the
EVTC header. Update the header parsing code to correctly handle the new
layout, and extract the revision. This will be used to parse the combat
events properly for both old and new versions of the .evtc files.

Also update the version of simpleArcParse to v1.0, and add a new
"revision" command which may be useful for debugging.

While at it, add a new test case for simpleArcParse to ensure that the
program can correctly read the new revision 1 files.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>