add module function to format and publish some encounters
Currently, we have the format-encounters.ps1 script for the purpose of
finding and formatting encounters to a discord web server.

This script is somewhat monolithic and currently relies heavily on the
gw2raidar API. This is problematic, as it is difficult to correlate
gw2raidar links back to the local evtc data we parsed.

It is difficult to extract portions of this logic for the
purpose of manually posting encounters, such as posting links to failed
runs.

Additionally, it's not easy to modify format-encounters to allow
publishing multiple success runs of the same boss. This means that if you
forget to run format-encounters, and happen to re-run a boss, you will
only upload the most recent kill.

This is mainly due to the nature of how format-encounters relies on
gw2raidar and does a backwards time lookup for finding the gw2raidar
links.

To pave the way for adding new ways of formatting scripts which doesn't
rely heavily on the gw2raidar API, add a new module function for
formatting and publishing an array of boss hash tables.

This function, Format-And-Publish-Some is heavily based on the
Format-And-Publish which already exists within the format-encounters.ps1
script.

This function will take an array of bosses, the guild which is publishing,
and the config object. Using this information it will generate a
PSCustomObject representing the webhook embed we wish to publish.

Careful note should be taken that the function *assumes* that all bosses
in the array were run on the same date, and were run by the same guild. It
looks this information up from the first element of the bosses array. This
assumption is useful so that each post has a header which includes the
guilds name and date, as well as a single footer with the players list.

Support for taking an arbitrary list of bosses and breaking it apart into
multiple publishes will be added in a future patch.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>