add a module function for converting accounts to discord pings
In format-encounters.ps1 there is logic to convert player account names
into their discord names by looking at guild configuration. This
functionality may be useful in other scripts as well as future module
functions.

Extract this logic and move it into the module. For now, we don't change
format-encounters.ps1, as the guild logic there is slightly different than
what we ultimately want in the module.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>