add a module function to create a $boss hashtable
The format-encounters.ps1 script essentially generates hashtables which
contain boss information. These boss hash tables are used by the
Format-And-Publish methods in order to convert the data into embeds for
the discord webhook API.

Currently, format-encounters.ps1 builds this data up through a combination
of gw2raidar lookups and reading the evtc files. The process for doing
this is heavily driven by the gw2raidar API, with looking up some of the
data from the local files as sort of an afterthought.

Future enhancements are going to move format and publish functions into
the module files, and we will provide multiple ways of determining which
bosses we want to publish at a time.

Add a new module function which can generate a boss hash table from only
the evtc local "extras" data generated by upload-logs.ps1

This function will lookup the evtc extras folder from the config, and try
to find the evtc data to build up a hash that is similar to (though not
exactly identical to) the boss hash tables currently in use by
format-encounters.

This includes storing the boss name, wing number, boss start time as a
Date object, the player list, the guild that ran the encounter, and the
dps.report link.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>