add a module function for publishing an arbitrary list of bosses
Add the Split-Bosses function which takes an array of bosses and splits it
apart into a hash of hashes.

The first layer of the hash is keyed by the date, and will contain a
sub-list of hash objects for the bosses.

The second hash, is keyed by guild name, and contains an array of bosses.
In this array are all the bosses which were run on the specific date, by
that specific guild.

Essentially, split the array of bosses into subsets each of which have the
same encounter date, and which are run by the same guild.

Use this function in the new Format-And-Publish-All funciton, which will
split apart an array of bosses and then call Format-And-Publish-Some on
each subset.

Additionally, support the "everything" parameter of guilds, in the same
way that format-encounters.ps1 does, by publishing each subset of bosses
first to its own guild, and then to every guild marked "everything" that
has a separate webhook URL.

This function very nearly re-creates the main logic from
format-encounters.

Future patches will re-work format-encounters.ps1 to build up an array of
boss hash tables, and then call this function. Additionally, a new script
which can take user input to determine an exact list of bosses to publish
could be added, so that users can easily re-publish specific boss
encounters, or even lookup and publish specific failed encounters, such as
during progression.

For now, these functions do not support gw2raidar. Support may be added to
them in the future, but it likely depends on better API support so that
looking up specific gw2raidar links becomes easier.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>