module: split bosses by encounter type as well
In addition to splitting the encounters by the guild and date, split them
apart by the type as well. This ensures that fractals will be published
separately from the raids.

Additionally, when formatting encounters, use the text "Wings" for raid
encounters, and "Fractals" for the fractal encounters.

In order to simplify the building of nested hashes, we add a Get-SubHash
function to get the sub element hash, which helps make the logic for
adding sub hash elements more clear.

Unfortunately, we cannot use this same behavior for the final array. This
is because of powershell not allowing array references to be modified.
Instead, the local variable would just get overwritten.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>