move log-output to the module file
Don't re-implement Log-Output in each script, but instead implement it in
the module file. To support this, add a new Set-Logfile function which
will set the global $script:logfile variable for that script run. This
way, callers can simply set the log file and then call Log-Output without
having to pass the file around everywhere.

This reduces code duplication between format-encounters and upload-logs,
and also enables use of Log-Output from within functions in the module
file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>