re-write format-encounters to use the new module functions
Completely re-write format encounters, dropping support for gw2raidar.
Now, we use the Format-And-Publish-All function available in the module
file.

Instead of trying to look up encounters via gw2raidar and then correlate
these back to the local dps.report data, we will just find all the
successful encounters since we last formatted.

This is faster, because we no longer have to scan the gw2raidar API.

We no longer show gw2raidar data, but this is ok, as the intention is to
drop support for gw2raidar until they provide a better API for finding the
URLs.

A future patch will modify the upload script to stop uploading encounters
to gw2raidar.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>