shorten "Conjured Amalgamate" display name
The "Conjured Amalgamate" name is so long that it forces the second column
embed over a little bit more than other names. This results in some
awkward spacing causing the second column to not line up properly.

Shorten the display of this name to "C. Amalgamate" so that the columns do
not become mis-aligned.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>