add configuration options for controlling upload
Add upload_dps_report and upload_gw2raidar configuration values, which can
be used to control whether encounters get uploaded to dps.report and
gw2raidar respectively.

If configured to "no" then no uploads will be done for that website. If
configured to "successful" then only successful encounters will be
uploaded. If configured to "all" then all encounters will be uploaded.

The values may be left unconfigured. If this is the case, then the default
for dps.report will be "successful", while the default for gw2raidar will
be "all". This is to preserve old behavior on configurations which have
not been updated to include the new values.

Add "validStrings" to the configuration fields, which allows specifying
exactly which strings are valid for a given configuration. This may be
useful to extend to other configuration fields in the future...

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>