re-add support for posting gw2raidar links
Add a function for downloading gw2raidar links, which generates a hash of
server start time to gw2raidar permalink. Add a function which takes a
list of bosses, and uses the generated gw2raidar link hash to correlate
the links to the encounter based on server start time. If the gw2raidar
link has an exact match, store the associated link as the correct
permalink for gw2raidar.

With this, when running format encounters, we can quickly obtain
permalinks without the insane overhead of the older system. It still isn't
*as* exact as a proper API for checking on specific uploads, but it seems
to work ok for now.

With this change, we have full support for gw2raidar and dps.report, and
we have the ability to enable or disable dps.report and gw2raidar
individually.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>