remove the unncessary gw2raidar_start_map data
The start_map data was originally used as a way of going from server start
time to an evtc data directory. This was necessary because
format-encounters *started* from gw2raidar data, and worked backwards to
the evtc data generated from local files.

Now that we have replaced the format-encounters implementation with one
which works forward starting from local data, the start map is no longer
necessary, since the server start time is directly stored within the evtc
extras folder.

Let's remove the start map creation, as it has been superseded by a better
method. Update the upload-logs.ps1 to print a message that it is safe to
remove the now unnecessary start map data and configuration variable.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>