format-encounters: use creation time instead of last write time
Using last write time for finding evtc extra directories is problematic,
because writes to new files inside the folders update the modified time.
This can cause accidental reposts by changing the modified time of an
object to later than the "new_format_time" which we save prior to starting
the filter.

The reason we save this time prior to searching is to avoid a race
condition where new evtc data is created after we search, but before we
post.

Instead, let's use creation time, which avoids this by relying on the
creation date of the directory. This should be more robust and avoid the
race condition mentioned above.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>