use full boss name when looking up the emoji
Recently, we added the ability to use a short name for some bosses, most
notably the Conjured Amalgamate. When looking up the emoji we used the
local $name variable. This now uses the shortname, which causes the boss
lookup to fail, returning no emoji text.

Fix this by directly using $boss.name instead, so that we look up the
correct emoji.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>