format: add the encounter duration to the discord report
Encounter duration is measured in seconds, so it's not as precise as the
duration recorded by the dps.report and gw2raidar sites, but it should be
suitable.

Extract the end time using the new simpleArcParse end_time command. Then
calculate the duration in seconds, and print it as "(minutes)m
(seconds)s".

This will be added to the end of the links to gw2raidar and dps.report,
since putting it on the boss name line could make the width too large.
Separate this time using a middledot unicode character.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>