don't use -Include with -LiteralPath
We use the -Include parameter when using Get-ChildItem in an attempt to
limit the searches to only include files which match a simple file glob.

Unfortunately this doesn't actually work, because we use -LiteralPath. In
fact, -Include seems to have zero effect in this case.

This appears to be due to the fact that -Include only works on -Path.

The functions appear to work today because almost no one has any *non
matching* paths within their arcdps.cbtlogs folder or within the
arcdps.uploadextras folder created by these scripts.

However, it is good practice to avoid finding files which don't match. We
fix this in two ways.

For format-encounters, we just switch to using -Filter, which is both
faster, and works with -LiteralPath.

For upload-logs.ps1 this is trickier, as we want to match multiple
extensions. In this case, we just move the check to a Where-Object filter
after the Get-ChildItem call. This shouldn't be significantly slower than
-Includes would have been, though it is a bit more verbose.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>