add a function to upload an old encounter
Add the Complete-Upload-To-DpsReport function. This function takes a $boss
hash table, and uploads the associated evtc.zip file to dps.report, if it
wasn't yet uploaded.

This is useful if the user wants to manually upload an encounter, and wil
lbe used by a future script which allows querying the upload extras
"database" and uploading specific encounters. Possibly this can be used to
upload specific failed encounters that you think may be useful or relevant
for your group.

One method of using this might be something like the following, from the
powershell console:

> Import-Module -Force -DisableNameChecking (Join-Path -Path "." -ChildPath l0g-101086.psm1)
> $config = Load-Configuration "l0g-101086-config.json" 2
> $boss = Load-From-EVTC $config "20181126-191247.evtc"
> Complete-UploadTo-DpsReport $config $boss

These steps will import the l0g-101086.psm1 module, load the
configuration, generate a boss object for the given EVTC file, and then
attempt to upload the EVTC file to dps.report.

This can be useful when attempting to upload an individual boss file that
was otherwise ignored by upload-logs.ps1

A future interactive publisher script is intended which will ease this
process so that users can interactively find EVTC files and upload them.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>