indicated failed encounters with "(FAILED)" string
Instead of showing the duration for failed encounters, show the string
"(FAILED"). This should only matter for encounters which are formatted
manually, but helps distinguish the failures from the successful
encounters.

Preferably we could change the boss name to red text, but discord markdown
doesn't seem to support something like this. Another alternative would be
a second emoji like a thumbs down or something, but that also seems too
vague.

For now, this "(FAILED)" should provide enough information.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>