simpleArcParse: stop using templates for cbtevent revisions
Replace the template-based code for handling evtc_cbtevent structure
differences.

This code was originally added in an attempt to reduce code duplication
when handling v0 and v2 events. The idea was to remove as much boiler
plate as possible. Unfortunately, we actually end up having boiler plate
at the top level of code in the main function, since we have to decide
which revision we're parsing.

This makes the main function rather annoying to read, and we also have to
litter the template type everywhere.

Replace this logic with a evtc_cbtevent class, which has a constructor to
create an evtc_cbtevent. This replaces the get_cbt_event_details function,
and allows us to drop the template bits.

We do end up still having some of that boiler plate code for extracting
the various fields, but now it's kept isolated within the evtc_cbtevent
class. Additionally this should let us do a better job if we ever have to
handle fields which aren't compatible between the two versions.:wq

To avoid duplicating a bunch of the boiler plate logic for extracting
cbtevent fields, add the CBTEVENT_ACCESSOR macro which will generate this
code at compile time.

Since this change should not produce any impact on the simpleArcParse
results (unless there are bugs!), increment only the patch version.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>