loosen version requirements of simpleArcParse
Going forward, I plan to keep the simpleArcParse version as a semantic
version scheme, based on the rules at https://semver.org/

This means that the version scheme will follow these rules:

Given a version number MAJOR.MINOR.PATCH, increment the:

MAJOR version when you make incompatible API changes,
MINOR version when you add functionality in a backwards-compatible manner,
and
PATCH version when you make backwards-compatible bug fixes.

With this in mind, we can loosen the strict requirements for the version
of simpleArcParse to the following:

(a) If the major versions are not equivalent then assume we cannot
possibly work with this version of simpleArcParse.

(b) even if the major versions are equivalent, only allow simpleArcParse
versions which are *at least* the minor version we expect. This assumes
that we will always bump the minor version when adding new (compatible)
features.

Add a new Check-SimpleArcParse-Version function to the module file which
will verify if the version string follows these rules.

It will return $true if the version is compatible, and $false otherwise.
Additionally, if it returns $false, then a message will be written to the
console indicating that the expected version is not compatible.

If the versions are compatible, but not what we expcted, then the function
will simply log to the output file instead, indicating that the version
does not match. If simpleArcParse is newer than expected, simply state
that the version is newer than the expected version.

If the *patch* version is older than expected, log a message that there is
an available update.

In this way, it should now be possible, (from this point on) to use newer
versions of simpleArcParse without breaking the scripts, as long as the
major version is not incremented.

This is primarily intended to ease development of the simpleArcParse
program, as new features can be added and the minor version updated,
without having to update the version in the upload-logs.ps1 file
constantly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>