support the new .zevtc file extension for compressed logs
Newer versions of ArcDPS since 20181212 now compress the EVTC files using
the new .zevtc file extension.

To support this, add two helper functions to the library module for
checking if a filename is an EVTC file, and for determining the
uncompressed name that we want to associate it with.

This simplifies the code in upload-logs.ps1 and avoids needing to further
duplicate some of the logic. Additionally this should allow future new
filename changes to be a bit easier to handle.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>