simpleArcParse: add support for more precise duration calculation
The current method of calculating duration is to take the server end time
from the CBTS_LOGEND and subtract the server start time from the
CBTS_LOGSTART events.

This works, but it's not very precise for a few reasons. First, it only
reports time in seconds. Second, it doesn't take into account the fact
that the logs can keep going for a couple seconds after the boss dies.
This results in "durations" being displayed which are longer than
expected.

Fix this by using the time field of the events to calculate durations in
millseconds.

Add the new "duration" option which will print the encounter duration
using this new method.

A future patch will update the upload script and have it actually extract
the duration.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>