simpleArcParse: fix bug for unknown ids
Using a "const char *" pointer for the boss_name works unless the boss id
is unknown. In this case, the code tried to create a string stream with
"Unknown encounter <id>", and then std::move that into the character
pointer.

This won't work. Instead, just change boss_name to be a string. Then we'll
automatically create a string when assigning it to a const C string, and
we'll be able to properly move the name in place.

Without this change, simpleArcParse will return garbage data for the name,
which can result in unexpected output. Ultimately, the format-encounters
script will crash on the garbage data.

Update the simpleArcParse version number to indicate a bug is fixed. Since
this doesn't change how callers should use the program, we only bump the
minor version.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>